requirements.txt
pygame-ce>=2.4.1
python-chess>=1.10.0
numpy>=1.26.0
